requirements.txt
--extra-index-url https://binary.alfabank.ru/artifactory/api/pypi/fs_etl-pypi/simple
pydantic==1.9
uvicorn>=0.15
fastapi==0.78
fastapi_utils==0.2
psycopg2-binary==2.9.3
python-dotenv==0.20
devtools==0.9.0
python-dateutil>=2.8
gitpython>=3.1.20
pyyaml==6.0.1
more_itertools==9.1.0
aiohttp==3.8.5
asyncssh==2.14.2
anyio==3.7.1

fs_common_lib==0.22.10
fs_db==0.17.13
